Fix facdb QAQC integer casting bug
Capacity is stored as a text/var char in facilities and unmapped
facilities, respectively. So it's something of a wild west.

A statement like `select sum(capacity::integer)...` is fine when
capacity is text representing an integer, but will fail with text
like '1.0'.

e.g.
`select '1'::integer;` is fine.
`select '1.0'::integer;` -> invalid input syntax for type integer: "1.0"

Perhaps there's a cleaner way to do this, but for now, I'm just casting
to a numeric, which is more permissive, then casting back to int after sum