Fixup work: move df sanitizer to geos. fns
In the prior flow, the dataset (dataframe) was sanitized by removing
dupes and changing np.nans -> None. It was done after the pipeline
function body was executed, but before the subsequent wrappers executed.

This inlines that sanitization into each function. It isn't pretty, but
it's explicit, and should mirror exactly what was happening before