Refactor facdb pipelines (remove decorator pattern) (#262)
Refactor facdb pipelines (remove decorator pattern) (#262)

* Inline @Prepare

* Inline @Export

* inline remaining decorators

* extract prepare and export steps to cli

* Fixup work: move df sanitizer to geos. fns

In the prior flow, the dataset (dataframe) was sanitized by removing
dupes and changing np.nans -> None. It was done after the pipeline
function body was executed, but before the subsequent wrappers executed.

This inlines that sanitization into each function. It isn't pretty, but
it's explicit, and should mirror exactly what was happening before

* Remove `None` default argument to pipeline fns

* Appease mypy

* Add mypy check for facilities db

* comment out kpdg tests for dcpy, docker, py

---------

Co-authored-by: Alex Richey <alexrichey@DCP-APPLE-2180.local>