fix presence of GeometryCollection in  `nyc_historic_buildings`
we were doing `ST_UNION` on a geometry with mixed types due to the `COALESCE`. this generated GeometryCollection values which aren't handled well by any subsequent queries or export files.

This also led to use buffering the lots and points of a multi-point landmark when only some points joined to a lot. Now we don't use the extra points if there are any lots.