Bytes Packaging With OTI_XLSX Generation (#1146)
* Move COLP single feature metadata package conftest

Slight refactor to move the COLP test package up a level
(out of lifecycle, and to the top-level of the tests)

* Assemble from BYTES w/ oti_xlsx generation

Add ability to assemble a dataset from Bytes, while generating the OTI
XLSX if a file of type oti_data_dictionary is specificed in the
metadata.

* Refactor: move code into lifecycle.package

lifecycle.assemble was an older incarnation of lifecycle.package, so
moving tests out. And lifecycle.distribute.bytes wasn't really doing
anything bytes specific. (At least not enough to warrant being separated
from the other packaging code)

No changes to code in the files. Just moving things around.

---------

Co-authored-by: Alex Richey <alexrichey@DCP-APPLE-2180.local>