More Metadata Changes (#1180)
* Add Typing for the Publishing Keyword Arguments

These arguments are being passed all over, and this enables some
type-safety. This is a potentially interesting use-case for
TypedDicts.

* Enable metadata only package+deployments.

As I go through LION+DCM+Zoning, I want to be able to assemble a package
of only metadata, and deploy only the metadata. For time considerations
more than anything

* Add remaining publication fields

* Use OrgMetadata in CLI tools

* Shuffle around the Org/Product/Dataset Overrides

Refactors the override classes, with the eventual goal of enabling
overrides/defaults from the org level. This is driven by the need to
set Org-level fields like `agency` at the top-level, and pass that along
to the Socrata connector.

* Enable Metadata at the Org Level

Allows us to set fields at the org-level, which can then be set at the
product-level, then dataset-level. This is mostly for fields like
agency, which needs to be passed to the Socrata publish connector

* Socrata Pub: pull new fields from metadata

Okay, now that we can set these fields at the org-level down through to
the dataset, we can remove the literals when we publish

* Add product metadata repo env var

It's really annoying to have to keep pasting the path to your metadata
repo. This adds the `PRODUCT_METADATA_REPO_PATH` env var, and makes that
path flag optional (it's now an override)

* Fix straggling kwarg errors

* Add test for md-only bytes pull, fix docs

* Throw an exception for missing socrata md fields

---------

Co-authored-by: Alex Richey <alexrichey@DCP-APPLE-2180.local>