PLUTO: tag tests based on PLUTO version type (major vs minor) (#1192)
* Add `version_type` env var in PLUTO recipes:

* Add dbt variable `version_type` from env var:
This dbt var is not used in code. Rather, we use the env var in a dbt CLI command. We add it to the file to check early that the env var exists (before creating tables and running tests)

* Minor tweak to dbt command

* Create new tag `major`

* Add test for PLUTO minor (same row count between current and previous PLUTO versions)

* Add `dcp-developments` to PLUTO minor recipe:
It is not actually needed for PLUTO minor build and tests. However, planning a recipe step fails without it

* Revise env var `version_strategy` setup