consolidate ingest source models
these models really are specific to the ingest template, not the
connector (the connector doesn't need to know it's own "type"). In the
connectors, I think it's easier to work with kwargs than create these
objects, whereas when we work from yaml (like in ingest) it's still
very helpful. So I think they fit better here.