remove check timestamps
I get why I did this -> very convenient to have a "date last checked"
in the config. But it's simple enough to use the ingest raw dataset
connector to check the "latest" archived timestamp instead, so I think
we should just go with that and keep immutability. Feels a little more
sane