add typed connector subregistries
If we have different apis for these connectors, at some points we
need to be able to declare what we expect to be able to do. Maybe this
is a little stringent right now (before getting a subregistry of our
default one, no api is available), so might need some reworking. I think
the utility of this is more shown when being used in ingest.