rework non/versioned connectors and their apis
Things just needed some cleaning. Get us to a point where we can use
the same api for these different "types" of connectors. Thank you
Alex for input on this. We likely have a bit of refining to go but
this worked well for the needs of this PR (see later when it's used
in ingest)