ingest - implement and make use of connector classes (#1519)
* remove check timestamps

I get why I did this -> very convenient to have a "date last checked"
in the config. But it's simple enough to use the ingest raw dataset
connector to check the "latest" archived timestamp instead, so I think
we should just go with that and keep immutability. Feels a little more
sane

* remove unused arg

* consolidate ingest source models

these models really are specific to the ingest template, not the
connector (the connector doesn't need to know it's own "type"). In the
connectors, I think it's easier to work with kwargs than create these
objects, whereas when we work from yaml (like in ingest) it's still
very helpful. So I think they fit better here.

* update ingest templates with changed type literal

* rework non/versioned connectors and their apis

Things just needed some cleaning. Get us to a point where we can use
the same api for these different "types" of connectors. Thank you
Alex for input on this. We likely have a bit of refining to go but
this worked well for the needs of this PR (see later when it's used
in ingest)

* add typed connector subregistries

If we have different apis for these connectors, at some points we
need to be able to declare what we expect to be able to do. Maybe this
is a little stringent right now (before getting a subregistry of our
default one, no api is available), so might need some reworking. I think
the utility of this is more shown when being used in ingest.

* add source connectors for ingest

Implement connector objects and tests for the various ingest sources

* add storage connectors, ingest datastore

two points here
- "ingest_datastore" connector that's decoupled from any specific
  storage. The idea is that user could set up storage for ingest with
  something like what's sort of outlined in #1555
- "storage" connectors underneath - I really needed a connector that
  both was non-versioned and also had some concept of  listing things
  after a prefix. This maybe will be scrapped in the future. Alex and I
  have talked about how maybe all connectors should  move towards
  "paths" (sort of RESTful) which would make this maybe redundant. But
  for now, I needed it.

* chk - ingest

* default push/pull implementation for versioned

reduce boilerplate. this could have gone either direction, but there
were cases where "pull_versioned" needed a specific implementation so
this felt more generalized.

* POST-REVIEW: deprecate 'Type'

* POST-REVIEW: rename 'drive'

* POST-REVIEW: just pass storage to ingest conn

* rework logic around existing ingest versions

* slightly different mypy hack for abstract generics