REWRITE!  loading of ACS data
I found everything about this module (from the name onwards) very
confusing. It's referenced in most of the remaining datasets, so I
want to make it usable before proceeding.

Most of the new functionality is in aggregate.load_aggregated.py, but
I also implemented it in two pipelines to give a sense of the changes.

Note: instead of changing anything that depends on the the indicator
code, I took the cowards way out, and just changed the reference to the
.run function on the aggregator.