CHANGES:  qol.assault_hospitalizations + ped_hosp
We didn't receive data for pedestrian hospitalizations, so I've
commented it out in the accessors.

For Assault Hospitalizations, the main change is that it was
calculating populations from a column that previously existed in the
assualt file. This is... not ideal.

Lucky for us, it was removed, so I've changed to use the Census populations.