CHANGES!  Get ACS economics working
Similar to demographics, we don't have a separate file anymore. Also...
There were columns in the source data that mapped to duplicate values
due to how we parse them. I don't have the energy to fix this, so I just
deleted the columns from the main tab, and added them to a tab called
dupes in the spreadsheet.

This also changes to actually filter for a specific set of columns from
the source. Otherwise we have too many columns in the output. But
comparing my build to builds from previous years, the total number of
columns has remained the same. So it appears that this doesn't change
the output at all.