CHANGES:  ho_prod.change_in_units
Hoo boy. Lots of changes here. This was previously pulling two Excel
files from the website: the first had the total number of houses in the
city by 2010 geos, then a crosswalk for that data to the pumas.

Instead, Winnie suggested using a version of the 2010 Census data with
2020 NTAs, maintained by Population. I trimmed it down to just the
HUnits column, and only the relevant geometries