Move old files to resources/_old/*
I'm leaving them in the repo because it's nice to have them as a reference for
what was done in the past, especially since for many of them, we didn't
receive data this year. It's quite confusing to have them sitting next
to fresh data though (and potentially a source of errors), so I'm moving them.