pluto 25v2 tweaks (#1687)
* update pluto FAR table

* use pts roll file for library archive

This file had a special file name
I verified that it had identical structure
to the usual pts file.

This commit is reverted just after this, including
for transparency in pr (this will be squashed)

* Revert "use pts roll file"

This reverts commit 994cd7dbca0ced89375765b5fd81fbd6d324e719.

* allow dev bucket for pluto input ghas

* autoformat pluto inputs ghas

* simplify pluto corrections

no more boilerplate in my house.

would love another pair of careful eyes - it seems
like all the individual sql files were doing the
exact same thing but I might have missed a special
case. lot area had a cast to numeric so I left it
as a special case

* pluto 25v2 corrections

* assign appbbls from unit lots if not found

see #1661 for more of a write-up on the root issue
decision was to just take the lowest unit appbbl
for any condos with associated units that were
missing a "prime" appbbl.

* remove histdist corrections

issue was fixed in source data

* handle null bbls in research