simplify types for socrata pulling
For one, the types here were originally pulled
from ingest.

Could have just moved them to `socrata`, but they
also felt slightly fussy - many of the functions
were only using some of the fields, and all of the
cases where the pydantic objects had to be created
just be to broken down again later felt a bit
clunky. Made sense in the context of file IO with
ingest templates, less so in our connector
philosophy.