ingest validation - harcode schema, use temp table
There's no need to have "build names" based on
branch - we're not trying to migrate the same
datasets in parallel. And this way you don't have
to worry about local env vars for running
comparisons locally

Also, for big tables, it's a little crazy to do
a big join for every single column to compare.
Could optimize slightly more (there are the double
outer joins) but this was the bigger offender.

Also adds option to load from s3 dev bucket. I
needed this for geosupport stuff which I didn't
feel like running on my comp since I'd need docker
as well as library which is the weirder process