migrate pluto inputs to ingest (#1937)
* ingest validation - harcode schema, use temp table

There's no need to have "build names" based on
branch - we're not trying to migrate the same
datasets in parallel. And this way you don't have
to worry about local env vars for running
comparisons locally

Also, for big tables, it's a little crazy to do
a big join for every single column to compare.
Could optimize slightly more (there are the double
outer joins) but this was the bigger offender.

Also adds option to load from s3 dev bucket. I
needed this for geosupport stuff which I didn't
feel like running on my comp since I'd need docker
as well as library which is the weirder process

* option to use dev image in pts gha

* fully custom image for ingest_single

* remove unused pluto mvt logic

This was a prototype/POC for some work with
AE, they went in a different direction (and we
never got over certain performance issues for the
overall workflow).

Seemed best just to scrap the unused code.

* simplify types for socrata pulling

For one, the types here were originally pulled
from ingest.

Could have just moved them to `socrata`, but they
also felt slightly fussy - many of the functions
were only using some of the fields, and all of the
cases where the pydantic objects had to be created
just be to broken down again later felt a bit
clunky. Made sense in the context of file IO with
ingest templates, less so in our connector
philosophy.

* clean up ingest template validation

There was just a little bit of weirdness around
where errors were getting raised versus objects
representing "violations" or validation errors
returned. This especially is needed for the next
commit, which adds functionality to the validation
step. Still could use a little refining I think.

* rework ingest source to be generic connector

To make it easier for custom/new sources, it
seemed worthwhile to make the source field of
the ingest template to be a little less finicky
and just work with registered connectors.

There's also a validation function that makes
sure that in whatever configuration dcpy is
running, that a specific template or all templates
have real, registered connectors as their sources
and have all the necessary arguments to `pull`
them

* tweaks to sftp semantics

just wanted to only call things "connectors" in
our connector code, and leave that language
out of the more pure utilities

* use sftp connector to pull from ginger

* copy in pluto geosupport code to dcpy

* pluto_pts template and logic

pts data validation:
https://github.com/NYCPlanning/data-engineering/pull/1937#issuecomment-3272034821

geocodes data validation:
https://github.com/NYCPlanning/data-engineering/pull/1937#issuecomment-3286552990

* pluto_input_numbldgs

This really should just happen in the pluto build.
One step at a time.

data validation:
https://github.com/NYCPlanning/data-engineering/pull/1937#issuecomment-3286454578

* somewhat in-progress cama template

cama data is not present on the server at the
moment, which makes it slightly hard to test!

* recurring ingest action for pluto_inputs, cleanup

run of action here:
https://github.com/NYCPlanning/data-engineering/actions/runs/17686449760