make new consolidated segment table
We have two needs of segments and their various
source tables
1. Tabular data, specific to the feature type
2. Geometry data, generally handled the same way
   across feature types

For the former, it makes sense to have staging
tables per feature type (index on segmentid).

For the latter, it's nice to have one stg/int (I
went with int because of some little complexity)
table that has rows from all source feature layers

This unified table has geom columns as well as
some basic identifying info that is either
duplicated or also in the geometry tables. This
table is meant to be the backbone of LION.