Move off recipes: Ingest and Library
Kicking the can here, to not interfere with ingest refactoring underway.

Some of this isn't strictly necessary. In particular, making the
ingest.connectors dynamic... I started down the path of fixing up the
unit tests, and it was made much more difficult that references to
`datastore` and `raw` connectors were set on import of the ingest
module. So I took a similar approach as in lifecycle.builds