LION - QOL, reworking of structure, and all segment source layers (#1979)
* disable a few tests

* tweak dat_loader command to make full output table

before I had just done this manually. Prod files
come at borough level, but it's easier for us
to compare at the citywide level

* clean up export script

* pin recipe and data source versions

for now, just an easy way to make sure all
versions are consistent

* cast curvature fields to bigint

not sure why this didn't happen to 25b, but some
values are too big

* remove unused model

* include row counts in poc comp

* simplify lgc logic

no need for two tables to rank the preferred ones
and nonpreferred ones separately and union

* make new consolidated segment table

We have two needs of segments and their various
source tables
1. Tabular data, specific to the feature type
2. Geometry data, generally handled the same way
   across feature types

For the former, it makes sense to have staging
tables per feature type (index on segmentid).

For the latter, it's nice to have one stg/int (I
went with int because of some little complexity)
table that has rows from all source feature layers

This unified table has geom columns as well as
some basic identifying info that is either
duplicated or also in the geometry tables. This
table is meant to be the backbone of LION.

* RENAME street and facecode from 2.2.1

* RENAME node-related transformations from 2.2.2

* RENAME offsets and polygons from 2.2.3, 2.2.13

as well as use the new segments table as source
rather than centerline

* RENAME/rework curves from 2.2.11

* RENAME/rework coincident segments from 2.6

* RENAME sedat folder from "8"

* RENAME other lion intermediate tables

* RENAME/reorganize product models

* rework scaffold of int__lion slightly

* int__lion simpler transformations

* coincident segments for non-centerline

* nsf seqnum part of the way there

* non-centerline zip codes