Refactor Recipes to Ingest Datastore and Cloudpath (#1939)
* Add Cloudpath and pytest-xdist requirements

* Move github tests to integration_tests and forbid callouts

* ignore mypy stub complaints for duckdb

If we use this more in the future, let's add stubs. For now...

* Delete recipe logging functionality

* Quality of life tweaks to dcpy and lifecycle confs

Adds some defaults to dcpy.configuration, and fixes pathing
problems for the Lifecycle data directory

* Move integration conftest up a dir level

These variables need to be set before anything from dcpy is imported.

* Add HybridPathedStorage and Integration tests

* CHANGES! Refactor ingest datastore to use pathedstorage

* change `edm.recipes` -> `edm.recipes.datasets`

* Move off recipes: source data versions in source_report

This isn't actually used. But it seems useful, and since it's
cross-lifecycle-stage, it seems like it fits in scripts

* Move off recipes: use default recipe conn in lifecycle

Also changes a few references away from recipes.* methods

* Move off recipes: qa source_report

Changes to just use data_loader

* Move off recipes: Ingest and Library

Kicking the can here, to not interfere with ingest refactoring underway.

Some of this isn't strictly necessary. In particular, making the
ingest.connectors dynamic... I started down the path of fixing up the
unit tests, and it was made much more difficult that references to
`datastore` and `raw` connectors were set on import of the ingest
module. So I took a similar approach as in lifecycle.builds

* Move off recipes: builds.test_plan

Use a mocked connector in builds.test_plan rather than patch

* Move off recipes: _dataset_type_from_ext

Please excuse the diff. Just staging the relevant bits (a big old hunk
was deleted from this file) to show precisely what changed

* Kill existing recipes unit tests

This is now integration tested

* Take a hacksaw to connectors.edm.recipes

* Fix ingest test_validate

* Change data_path to use os.getcwd()

* post-review: snip lingering recipes.Dataset(...)

* post-review: expand on test-skipping message