first pass at 'flipping' fields
So with the flipped geoms in commit 1, protosegments go through the
geospatial transformations like any other segment fully realized.
However, some of these fields are more tabular things that still need
special rules if we're on a borough boundary or something like that.

(i.e. we have a protosegment that's on a boro boundary that gets its
left sanitation subsection from its centerline segment but one of them
needs to be made null because it's not supposed to be defined if that
side of the segment is another boro)

We also needed some data from the stg__altsegmentdata_proto table for
proto segments, very similar to how we join in other staging tables
for lookup info for specific feature classes